% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auto_factorize}
\alias{auto_factorize}
\title{Convert character columns to factors}
\usage{
auto_factorize(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
The data frame with character columns converted to factors.
}
\description{
Converts all character columns in a data frame to factors, replicating
the behaviour of \code{\link[stats]{lm}} and \code{\link[stats]{glm}}.
}
\keyword{internal}
