% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_val}
\alias{default_val}
\title{Default value for a variable}
\usage{
default_val(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The default (central) value.
}
\description{
Returns the median for numeric variables and the mode for factors,
characters, and logicals.
}
\keyword{internal}
