% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{from_zelig2_model}
\alias{from_zelig2_model}
\title{Extract the Fitted Model from a zelig2 Object}
\usage{
from_zelig2_model(object)
}
\arguments{
\item{object}{A \code{zelig2} object.}
}
\value{
The underlying fitted model object.
}
\description{
Returns the underlying model object (e.g., a \code{glm} or \code{fixest}
fit) stored inside a \code{zelig2} object. This is useful for calling
methods that are specific to the fitted-model class.
}
\examples{
data(mtcars)
z <- zelig2(mpg ~ hp + wt, model = "ls", data = mtcars)
fit <- from_zelig2_model(z)
class(fit)  # "glm" "lm"

}
\seealso{
\code{\link{to_zelig2}}.
}
