% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{get_model_spec}
\alias{get_model_spec}
\title{Retrieve a model specification}
\usage{
get_model_spec(name)
}
\arguments{
\item{name}{Character string model name.}
}
\value{
A named list with the model specification.
}
\description{
Retrieve a model specification
}
\keyword{internal}
