% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest.R
\name{has_fixed_effects}
\alias{has_fixed_effects}
\title{Check whether a formula contains fixed effects}
\usage{
has_fixed_effects(formula, fixef = NULL)
}
\arguments{
\item{formula}{A model formula.}

\item{fixef}{Optional fixed-effects specification.}
}
\value{
Logical.
}
\description{
Check whether a formula contains fixed effects
}
\keyword{internal}
