% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{list_models}
\alias{list_models}
\title{List Available Models}
\usage{
list_models()
}
\value{
A character vector of model names.
}
\description{
Returns the names of all model families registered in the \pkg{zelig2}
model registry.
}
\examples{
list_models()

}
