% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mode_zelig2}
\alias{mode_zelig2}
\title{Statistical mode}
\usage{
mode_zelig2(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The most frequent value.
}
\description{
Returns the most frequent value in a vector. For ties, returns the first
value encountered.
}
\keyword{internal}
