% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qi-helpers.R
\name{qi-helpers}
\alias{qi-helpers}
\title{Quantity of Interest Helper Functions}
\description{
Low-level functions for drawing parameters from their sampling distribution
and computing expected values (EV) and predicted values (PV) for each
model family.
}
\keyword{internal}
