% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{resolve_weights}
\alias{resolve_weights}
\title{Resolve weights to a formula}
\usage{
resolve_weights(weights, data)
}
\arguments{
\item{weights}{Weight specification (formula, string, or numeric vector).}

\item{data}{A data frame.}
}
\value{
A one-sided formula.
}
\description{
Resolve weights to a formula
}
\keyword{internal}
