% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{to_formula}
\alias{to_formula}
\title{Convert to a one-sided formula}
\usage{
to_formula(x)
}
\arguments{
\item{x}{A formula or character string.}
}
\value{
A one-sided formula.
}
\description{
Convert to a one-sided formula
}
\keyword{internal}
