% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{to_zelig2}
\alias{to_zelig2}
\title{Convert a Fitted Model to a zelig2 Object}
\usage{
to_zelig2(fit, model, data, vcov_type = "default", cluster = NULL, num = 1000L)
}
\arguments{
\item{fit}{A fitted model object (e.g., class \code{glm}, \code{lm},
\code{negbin}, \code{svyglm}, \code{fixest}).}

\item{model}{Character string naming the \pkg{zelig2} model type that
corresponds to \code{fit} (e.g., \code{"ls"}, \code{"logit"}).}

\item{data}{The data frame that was used to fit \code{fit}.}

\item{vcov_type}{Variance--covariance estimator (default \code{"default"}).
See \code{\link{zelig2}}.}

\item{cluster}{Cluster variable for clustered SEs.}

\item{num}{Integer; number of simulation draws (default 1000).}
}
\value{
A \code{zelig2} object.
}
\description{
Wraps an existing fitted model (e.g., from \code{\link[stats]{glm}} or
\code{\link[survey]{svyglm}}) in the \code{zelig2} structure so that
\code{\link{setx}}, \code{\link{sim}}, and \code{\link{plot.zelig2}} can be
applied to it.
}
\examples{
data(mtcars)
raw_fit <- glm(mpg ~ hp + wt, data = mtcars, family = gaussian())
z <- to_zelig2(raw_fit, model = "ls", data = mtcars)
z <- setx(z, hp = 150) |> sim()
summary(z)

}
\seealso{
\code{\link{zelig2}}, \code{\link{from_zelig2_model}}.
}
