% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{vcov.zelig2}
\alias{vcov.zelig2}
\title{Extract the Variance--Covariance Matrix}
\usage{
\method{vcov}{zelig2}(object, ...)
}
\arguments{
\item{object}{A \code{zelig2} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A square numeric matrix.
}
\description{
Returns the variance--covariance matrix of the estimated coefficients,
using whatever estimator was specified via \code{vcov_type} in
\code{\link{zelig2}}.
}
\examples{
data(mtcars)
z <- zelig2(mpg ~ hp + wt, model = "ls", data = mtcars,
            vcov_type = "robust")
vcov(z)

}
